+++ CHANGES.rst
@@ -32,6 +32,11 @@ Changelog
 - `BaseDGHV.printXhtml` `clean` parameter is now `True` by default so it will
   call `separate_images` to avoid several `<img>` in same `<p>`.
   [gbastien]
+- When an error occurs on the `MeetingConfig` because of a field in a fieldset
+  that is not currently viewable we get a validation error but we do not
+  know why.  Display every validation errors at the top of the page so the user
+  see what is happening.
+  [gbastien]
 
 4.2b7 (2020-12-08)
 ------------------